Create function to update winner div content
The winner div did not change when a winner won, so a function was created to make its text content change. The loop would execute again, creating a new unnecessary array, when logging that text content change, so an additional array was created to only make the loop execute once.